+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.9.6
+* New platform: ASR920 support
+* New platform: ISR 4331 support
+* New feature: Control creds in config.yaml
+* Bug fix: Unrequested debug output
+* Bug fix: Several issues in `schedule_reload()`
+* Refactor: move connection to `open()` method
+
 ## 0.9.5
 * Bug fix: `reload_at` variable ignored
 * Bug fix: Disable reload ignored